+++ Naive/src/StepNaive.h
@@ -1,21 +1,19 @@
 #ifndef STEPNAIVE_H
 #define STEPNAIVE_H
 
-#include "Step.h"
-#include <algorithm>
 #include <cmath>
 #include <iterator>
+#include "Step.h"
 
-class StepNaive : public Step {
-  std::vector<double> forces;
-  std::vector<double> velocities;
-  std::vector<double> weights;
-
+class StepNaive  : public Step {
+    std::vector<float> forces;
+    std::vector<float> velocities;
+    std::vector<float> weights;
 public:
-  StepNaive(std::vector<double> &masses, unsigned N);
-  ~StepNaive();
-  void initialize();
-  void compute(tf3 &positions, double dt);
+    StepNaive(std::vector<float>& masses, unsigned N);
+    ~StepNaive();
+    void initialize();
+    void compute(tf3& positions, float dt);
 };
 
 #endif
