+++ Naive/src/StepNaiveCUDA.h
@@ -5,20 +5,21 @@
 
 #include "RandomGenerators.h"
 
-#include "ComputationsCuda.h"
 #include "Step.h"
+#include "ComputationsCuda.h"
 
 class StepNaiveCuda : public Step {
-  Computations *c;
-  RandomGenerators *rg;
-  thrust::host_vector<double> weights;
-  thrust::host_vector<double> velocities;
-
+  Computations* c;
+  RandomGenerators* rg;
+  thrust::host_vector<float> weights;
+  thrust::host_vector<float> velocities;
 public:
-  StepNaiveCuda(std::vector<double> masses, unsigned N);
-  ~StepNaiveCuda();
-  void initialize() {}
-  void compute(tf3 &positions, double dt);
+    StepNaiveCuda(std::vector<float> masses, unsigned N);
+    ~StepNaiveCuda();
+    void initialize() {
+
+    }
+    void compute(tf3& positions, float dt);
 };
 
 #endif
